requirements_server.txt
fastapi==0.110.1
uvicorn==0.29.0
python-multipart==0.0.9
pyannote.audio==3.1.1
faster-whisper==1.0.1
google-genai>=1.0.0
torch
torchaudio
